Added a status bar property
By manipulating this property you can change the offset the size of the AKTabBarView.
Reduces the height of the content view by n
Increases the y value of the content view by n
increases the y value of the tabbar by n

Essentially this enables you to restore the iOS 6 behaviour and appearance of the status bar!